work_with_database/requirements.txt
django
psycopg2-binary
python-slugify
